refactor: move and mutualism properties components
