+++ gravitee-apim-console-webui/src/management/api/apis-routing.module.ts
@@ -60,10 +60,8 @@ import { ApiCreationGetStartedComponent } from './creation-get-started/api-creat
 import { ApiCreationV4Component } from './creation-v4/api-creation-v4.component';
 import { ApiCreationV4ConfirmationComponent } from './creation-v4/api-creation-v4-confirmation.component';
 import { ApiCreationV2Component } from './creation-v2/steps/api-creation-v2.component';
-import { ApiPropertiesComponent } from './proxy/properties/properties/api-properties.component';
 import { ApiDocumentationV4Component } from './documentation-v4/api-documentation-v4.component';
 import { ApiDocumentationV4EditPageComponent } from './documentation-v4/documentation-edit-page/api-documentation-v4-edit-page.component';
-import { ApiDynamicPropertiesComponent } from './proxy/properties/dynamic-properties/api-dynamic-properties.component';
 import { ApiRuntimeLogsDetailsComponent } from './api-traffic-v4/runtime-logs-details/api-runtime-logs-details.component';
 import { HasApiPermissionGuard } from './has-api-permission.guard';
 import { GioPolicyStudioLayoutComponent } from './policy-studio-v2/gio-policy-studio-layout.component';
@@ -76,8 +74,10 @@ import { ApiProxyEndpointListComponent } from './endpoints/list/api-proxy-endpoi
 import { ApiGeneralMembersComponent } from './user-group-access/members/api-general-members.component';
 import { ApiAuditListComponent } from './api-audit-list/api-audit-list.component';
 import { ApiAuditLogsComponent } from './api-audit-logs/api-audit-logs.component';
-import { ApiDynamicPropertiesV4Component } from './properties-v4/dynamic-properties/api-dynamic-properties-v4.component';
+import { ApiDynamicPropertiesV4Component } from './properties/components/dynamic-properties-v4/api-dynamic-properties-v4.component';
 import { ApiDeploymentConfigurationComponent } from './deployment-configuration-v4/api-deployment-configuration.component';
+import { ApiDynamicPropertiesComponent } from './properties/components/dynamic-properties-v2/api-dynamic-properties.component';
+import { ApiPropertiesComponent } from './properties/properties/api-properties.component';
 import { ApiNotificationComponent } from './api-notification/api-notification.component';
 
 import { DocumentationManagementComponent } from '../../components/documentation/documentation-management.component';
