+++ gravitee-apim-console-webui/src/management/api/properties/components/dynamic-properties-v2/api-dynamic-properties-v2.module.ts
@@ -62,4 +62,4 @@ import { ApiDynamicPropertiesComponent } from './api-dynamic-properties.componen
   declarations: [ApiDynamicPropertiesComponent],
   exports: [ApiDynamicPropertiesComponent],
 })
-export class ApiDynamicPropertiesModule {}
+export class ApiDynamicPropertiesV2Module {}
