+++ gravitee-apim-console-webui/src/management/api/properties/components/dynamic-properties-v2/api-dynamic-properties.component.html
