+++ gravitee-apim-console-webui/src/management/api/properties/components/dynamic-properties-v2/api-dynamic-properties.component.scss
@@ -1,8 +1,8 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
-@use '../../../../scss/gio-layout' as gio-layout;
+@use '../../../../../scss/gio-layout' as gio-layout;
 
 $typography: map.get(gio.$mat-legacy-theme, typography);
