+++ gravitee-apim-console-webui/src/management/api/properties/components/dynamic-properties-v4/api-dynamic-properties-v4.component.scss
@@ -1,6 +1,6 @@
 @use 'sass:map';
-@use '@angular/material' as mat;
-@use '@gravitee/ui-particles-angular' as gio;
+@use '../../../../../../node_modules/@angular/material/index' as mat;
+@use '../../../../../../node_modules/@gravitee/ui-particles-angular/index' as gio;
 
 @use '../../../../../scss/gio-layout' as gio-layout;
