+++ gravitee-apim-console-webui/src/management/api/properties/properties/api-properties.component.html
