+++ gravitee-apim-console-webui/src/management/api/properties/properties/properties-add-dialog/properties-add-dialog.harness.ts
